+++ src/app/sidebar/SidebarLinks.ts
@@ -109,7 +109,12 @@ export const sidebarLinks = {
     },
     other: {
         label: OVERVIEW_SECTIONS.other.title,
-        links: [getSectionLinkItem(SECTIONS_MAP.programDisaggregation)],
+        links: [
+            getOverviewLinkItem(OVERVIEW_SECTIONS.other),
+            getSectionLinkItem(SECTIONS_MAP.programDisaggregation),
+            getSectionLinkItem(SECTIONS_MAP.optionGroup),
+            getSectionLinkItem(SECTIONS_MAP.optionGroupSet),
+        ],
     },
 } satisfies SidebarLinks