+++ src/lib/constants/sections.ts
@@ -255,6 +255,13 @@ export const SCHEMA_SECTIONS = {
         titlePlural: i18n.t('Constants'),
         parentSectionKey: 'other',
     },
+    optionGroup: {
+        name: SchemaName.optionGroup,
+        namePlural: 'optionGroups',
+        title: i18n.t('Option group'),
+        titlePlural: i18n.t('Option groups'),
+        parentSectionKey: 'other',
+    },
     attribute: {
         name: SchemaName.attribute,
         namePlural: 'attributes',
@@ -267,6 +274,13 @@ export const SCHEMA_SECTIONS = {
         namePlural: 'optionSets',
         title: i18n.t('Option set'),
         titlePlural: i18n.t('Option sets'),
+        parentSectionKey: 'other',
+    },
+    optionGroupSet: {
+        name: SchemaName.optionGroupSet,
+        namePlural: 'optionGroupSets',
+        title: i18n.t('Option group set'),
+        titlePlural: i18n.t('Option group sets'),
         parentSectionKey: 'other',
     },
     predictor: {
@@ -373,7 +387,7 @@ export const OVERVIEW_SECTIONS = {
     },
     other: {
         name: 'other',
-        namePlural: 'others',
+        namePlural: 'other',
         title: i18n.t('Other'),
         titlePlural: i18n.t('Other'),
         componentName: 'Other',