+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+# [0.61.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.60.2...v0.61.0) (2025-08-20)
+
+
+### Features
+
+* add option group sets [DHIS2-19841] [DHIS2-19840] ([#616](https://github.com/dhis2/maintenance-app-beta/issues/616)) ([679b5d8](https://github.com/dhis2/maintenance-app-beta/commit/679b5d8c85463a715d850a3b960834e3cc62404b))
+
+## [0.60.2](https://github.com/dhis2/maintenance-app-beta/compare/v0.60.1...v0.60.2) (2025-08-18)
+
+
+### Bug Fixes
+
+* add missing scrolling hooks ([#619](https://github.com/dhis2/maintenance-app-beta/issues/619)) ([a0a9d23](https://github.com/dhis2/maintenance-app-beta/commit/a0a9d23affdb767797bc2ae4b598fe1e73907a65))
+* change text of buttons to be consistent with other soft updates ([#618](https://github.com/dhis2/maintenance-app-beta/issues/618)) [skip release] ([e057e7e](https://github.com/dhis2/maintenance-app-beta/commit/e057e7e5c2f897af23058f17cffbdba460e0de87))
+
+## [0.60.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.60.0...v0.60.1) (2025-08-18)
+
+
+### Bug Fixes
+
+* default to horizontal tabs when choosing render as tabs ([#617](https://github.com/dhis2/maintenance-app-beta/issues/617)) ([2d22615](https://github.com/dhis2/maintenance-app-beta/commit/2d22615ac0444b7d032887eeefe6a03b13e733f6))
+
+# [0.60.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.59.0...v0.60.0) (2025-08-14)
+
+
+### Features
+
+* add option group list ([#614](https://github.com/dhis2/maintenance-app-beta/issues/614)) ([3aab748](https://github.com/dhis2/maintenance-app-beta/commit/3aab74861b04111f887696cf018c5ff492cf5cdc))
+
+# [0.59.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.58.0...v0.59.0) (2025-08-13)
+
+
+### Features
+
+* add reordering sections modal ([#609](https://github.com/dhis2/maintenance-app-beta/issues/609)) ([6daab39](https://github.com/dhis2/maintenance-app-beta/commit/6daab391ceba8eeb4d607b202818402fc4c83415))
+* definition and edit of custom forms [DHIS2-19090] ([#603](https://github.com/dhis2/maintenance-app-beta/issues/603)) ([6e23378](https://github.com/dhis2/maintenance-app-beta/commit/6e23378d2161283edae9b2cf1473680688403931))
+
+# [0.58.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.57.0...v0.58.0) (2025-08-12)
+
+
+### Features
+
+* sections display options ([#607](https://github.com/dhis2/maintenance-app-beta/issues/607)) ([3b8203a](https://github.com/dhis2/maintenance-app-beta/commit/3b8203a176f5f70769a8c0e36be4716da30829f7))
+
 # [0.57.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.56.2...v0.57.0) (2025-08-04)